CHORE: ImageSeletButton을 ProfileImageView로 변경(#74)
- ImageSelectButton을 ProfileImageView로 변경
- Button의 addTarget을 UITapGestureRecognizer의 addGestureRecognizer로 변경
